fix: Controllers methods, Query methods , Service methods

+++ server/src/main/java/ewm/category/service/CategoryService.java
@@ -4,11 +4,13 @@ import ewm.category.Category;
 import ewm.category.dto.CategoryDto;
 import ewm.category.repository.CategoryRepository;
 import ewm.exceptions.ObjectNotFoundException;
+import ewm.user.service.UserService;
 import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -17,6 +19,7 @@ import java.util.stream.Collectors;
 @AllArgsConstructor
 public class CategoryService {
     private final CategoryRepository categoryRepository;
+    private final UserService userService;
     private final ModelMapper modelMapper;
 
     public List<CategoryDto> getAllCategory(int from, int size) {
@@ -31,20 +34,20 @@ public class CategoryService {
                 .orElseThrow(() -> new ObjectNotFoundException("Категория не найдена")));
     }
 
+    @Transactional
     public CategoryDto createCategory(CategoryDto categoryDto) {
         return toCategoryDto(categoryRepository.save(toCategory(categoryDto)));
     }
 
+    @Transactional
     public CategoryDto updateCategory(CategoryDto categoryDto) {
         return toCategoryDto(categoryRepository.save(toCategory(categoryDto)));
     }
 
+    @Transactional
     public void deleteCategoryById(int id) {
-        try {
-            categoryRepository.deleteById(id);
-        } catch (NullPointerException e) {
-            throw new ObjectNotFoundException("Категория не найдена");
-        }
+        if(categoryRepository.existsById(id)) categoryRepository.deleteById(id);
+        else throw new ObjectNotFoundException("Категория не найдена");
     }
 
     private Category toCategory(CategoryDto categoryDto) {
@@ -53,5 +56,10 @@ public class CategoryService {
 
     private CategoryDto toCategoryDto(Category category) {
         return modelMapper.map(category, CategoryDto.class);
+    }
+    private void validateUser(Integer userId) {
+        if (!userService.isUserCreated(userId)) {
+            throw new ObjectNotFoundException("User is not created");
+        }
     }
 }
